Fix runtime errors. Now works on Windows with mcp2221 library.

+++ lib/i2c_interface.h
@@ -5,5 +5,5 @@
 typedef enum { ST_OK, ST_FAIL } STAT;
 
 STAT i2c_init();
-STAT i2c_write(uint8_t addr, uint8_t* buffer, uint8_t len);
+STAT i2c_write(uint8_t addr, uint8_t* buffer, uint16_t len);
 STAT i2c_deinit();
